new dict api

+++ game_scene.py
@@ -77,44 +77,49 @@ class MainGame(Entity):
         self.started = True
         word = self.phoneme_store.get_new_word()
         if word is not None:
-            self.update_counter = 0
-            self.help_text.text = f'The word is: "{self.phoneme_store.word}"\nLeft click in the green area to lay a phoneme,\nright click to pick one up.'
-            self.help_text.enable()
-            self.score_text.text = f'Score: {self.score}'
-            self.score_text.enable()
-            if self.voxels:
-                self.destroy_all()
-                self.help_text.text = self.phoneme_store.word
-                self.correct = False
-            self.create_clouds()
-            for z in range(ARENA_DEPTH + 1):
-                voxel_side_wall = Voxel(self.phoneme_store, self, position=(-1, 1, z))
-                voxel_other_side_wall = Voxel(self.phoneme_store, self, position=(len(self.phoneme_store.phonemes), 1, z))
-                self.voxels.append(voxel_side_wall)
-                self.voxels.append(voxel_other_side_wall)
-                for x in range(len(self.phoneme_store.phonemes)):
-                    voxel = Voxel(self.phoneme_store, self, position=(x, 0, z))
-                    voxel_wall = Voxel(self.phoneme_store, self, position=(x, 1, -1))
-                    self.voxels.append(voxel)
-                    self.voxels.append(voxel_wall)
-                    if voxel.position[2] == ARENA_DEPTH:
-                        voxel.color = color.lime
-                        voxel.texture = 'white_cube'
-            if self.player:
-                self.player.y = 0
-                self.player.x = len(self.phoneme_store.phonemes) // 2
-                self.player.z = 1
+            if self.phoneme_store.pron() is not None:
+                self.update_counter = 0
+                self.help_text.text = f'The word is: "{self.phoneme_store.word}"\nLeft click in the green area to lay a phoneme,\nright click to pick one up.'
+                self.help_text.enable()
+                self.score_text.text = f'Score: {self.score}'
+                self.score_text.enable()
+                if self.voxels:
+                    self.destroy_all()
+                    self.help_text.text = self.phoneme_store.word
+                    self.correct = False
+                for z in range(ARENA_DEPTH + 1):
+                    voxel_side_wall = Voxel(self.phoneme_store, self, position=(-1, 1, z))
+                    voxel_other_side_wall = Voxel(self.phoneme_store, self, position=(len(self.phoneme_store.phonemes), 1, z))
+                    self.voxels.append(voxel_side_wall)
+                    self.voxels.append(voxel_other_side_wall)
+                    for x in range(len(self.phoneme_store.phonemes)):
+                        voxel = Voxel(self.phoneme_store, self, position=(x, 0, z))
+                        voxel_wall = Voxel(self.phoneme_store, self, position=(x, 1, -1))
+                        self.voxels.append(voxel)
+                        self.voxels.append(voxel_wall)
+                        if voxel.position[2] == ARENA_DEPTH:
+                            voxel.color = color.lime
+                            voxel.texture = 'white_cube'
+                self.create_clouds()
+                if self.player:
+                    self.player.y = 0
+                    self.player.x = len(self.phoneme_store.phonemes) // 2
+                    self.player.z = 1
+                else:
+                    self.generate_player()
             else:
-                self.generate_player()
+                return self.build()
         else:
             self.end_game()
 
 
     def build_platform(self):
+
         self.ground.enable()
         self.update_counter = None
         self.reset_text.enable()
         self.give_up_button.disable()
+        self.create_clouds()
         self.started = False
 
     def update_score(self):
@@ -140,14 +145,18 @@ class MainGame(Entity):
         self.next_block.rotation = Vec3(10, self.rotated_y, 30)
 
     def update(self):
+        print('working')
         if self.player.y <= -100:
             self.player.y = 0
             self.score -= 1
         self.update_score()
-        self.next_block.texture = 'index'
         if self.phoneme_store:
+            print('Phoneme store:')
             if self.phoneme_store.phonemes:
-                self.next_block.texture = self.phoneme_store.textures[self.phoneme_store.phonemes[-1]]
+                print(self.phoneme_store.phonemes)
+                self.next_block.texture = PhonemeEngine.textures[self.phoneme_store.phonemes[-1]]
+        else:
+            self.next_block.texture = 'index'
         self.spin_block()
         if self.update_counter is not None:
             self.update_counter += 1
